+++ llvm-hs/src/LLVM/Internal/FastMathFlags.hs
@@ -21,15 +21,16 @@ import LLVM.Internal.EncodeAST
 import qualified LLVM.AST as A
 
 instance EncodeM IO A.FastMathFlags FFI.FastMathFlags where
-  encodeM A.NoFastMathFlags = return 0
-  encodeM A.UnsafeAlgebra = return FFI.fastMathFlagsUnsafeAlgebra
-  encodeM f = return $ foldr1 (.|.) [ 
+  encodeM f = return $ foldr1 (.|.) [
                if a f then b else 0
                | (a,b) <- [
+                (A.allowReassoc, FFI.fastMathFlagsAllowReassoc),
                 (A.noNaNs, FFI.fastMathFlagsNoNaNs),
                 (A.noInfs, FFI.fastMathFlagsNoInfs),
                 (A.noSignedZeros, FFI.fastMathFlagsNoSignedZeros),
-                (A.allowReciprocal, FFI.fastMathFlagsAllowReciprocal)
+                (A.allowReciprocal, FFI.fastMathFlagsAllowReciprocal),
+                (A.allowContract, FFI.fastMathFlagsAllowContract),
+                (A.approxFunc, FFI.fastMathFlagsApproxFunc)
                ] 
               ]
 
@@ -40,12 +41,13 @@ instance EncodeM EncodeAST A.FastMathFlags () where
     anyContToM $ bracket (FFI.setFastMathFlags builder f) (\() -> FFI.setFastMathFlags builder 0)
 
 instance Monad m => DecodeM m A.FastMathFlags FFI.FastMathFlags where
-  decodeM 0 = return A.NoFastMathFlags
-  decodeM f | FFI.fastMathFlagsUnsafeAlgebra .&. f /= 0 = return A.UnsafeAlgebra
   decodeM f = return A.FastMathFlags {
+                A.allowReassoc = FFI.fastMathFlagsAllowReassoc .&. f /= 0,
                 A.noNaNs = FFI.fastMathFlagsNoNaNs .&. f /= 0,
                 A.noInfs = FFI.fastMathFlagsNoInfs .&. f /= 0,
                 A.noSignedZeros = FFI.fastMathFlagsNoSignedZeros .&. f /= 0,
-                A.allowReciprocal = FFI.fastMathFlagsAllowReciprocal .&. f /= 0
+                A.allowReciprocal = FFI.fastMathFlagsAllowReciprocal .&. f /= 0,
+                A.allowContract = FFI.fastMathFlagsAllowContract .&. f /= 0,
+                A.approxFunc = FFI.fastMathFlagsApproxFunc .&. f /= 0
               }